requirements.txt
pandas==2.2.1
lifelines==0.29.0
scipy
numpy==1.26.4
matplotlib==3.8.0
seaborn==0.12.2
tqdm==4.66.5
miceforest==6.0.3
plotnine==0.13.6
rpy2==3.5.16
pytest==8.3.2
lxml==5.2.1
pytest==8.3.2
scikit-survival==0.23.0
plotnine==0.13.6
notebook==7.0.3
html5lib==1.1
pyYAML
